+++ src/main/java/lotto/generator/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.generator;
 
 import java.util.List;
 import lotto.domain.Lotto;
