+++ src/main/java/lotto/generator/RandomLottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
