+++ src/test/java/lotto/WinningLottoNumbersTest.java
@@ -4,6 +4,9 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import java.util.List;
+import lotto.domain.BonusNumber;
+import lotto.domain.WinningLottoNumbers;
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
